inventario-backend/requirements.txt
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-OIDC==2.2.2
PyJwt==2.10.1
flask-marshmallow==1.2.1
psycopg2-binary==2.9.10
marshmallow-sqlalchemy==1.1.0
gunicorn==23.0.0
setuptools
pydantic

# dev/tests
pytest
pytest-icdiff
mypy
pylint
requests
tenacity